A sketchy sketch of directory and file manipulation
To complete the sketch we need a bunch of new small algebras,
wrappers for option data types, and a pluggable abstraction over
streaming libraries. TODOs cover all the missing pieces.

I've opted to break the algebra into read only, write only, and a
unified super type. I thought about breaking things down a lot more, but
was too much design for something that's really just trying to wrap an
existing API.

With the streaming abstraction and good vertical integration of sub
algebras working with open buffers and handles, this starts to feel like
a rather high level API, but it's not. It needs to preserve all the
rambling disorganized power of the java.nio.file.Files interface
because that's the point. Higher level abstractions will be MUCH less
powerful and are a separate problem.