Implement NIO Double buffer
First of the buffer family to be implemented. There was some major
refactoring to get this together. Especially regarding the putBuffer
method which had to be reimplemented in terms of array copy operations
in cases where differing implementations of BufferIO are combined. In
practice this is likely to be rare, but since the type system doesn't
rule it out it had to be handled.