Modify BufferIO modules to tighten typing
ByteBufferIO and ByteBufferReadIO implementation types (i.e.
NIOByteBufferReadIO and NIOByteBufferIO) need to be available on the
algebras returned by the ByteBufferIO.Module. In writing FileChannel
algebras we must be able to specify specific families of ByteBuffer
implementations as a type parameter on the algebra itself to ensure
alignment.

In other words an NIOSeekableChannelReadIO's read method must be able to
require an NIOByteBuffer (as opposed to a FreeByteBuffer for instance).

Because client code may be able to know it'll have an NIOByteBufferIO
instead of a generic ByteBufferIO I needed to make the val member
holding the actual mutable buffer package private.