Add IsoMutable instances for NIOByteChannelIOs
So it's kinda ugly, but SeekableByteChannel won't tell you if it's open
for just reading or reading and writing. If a client sticks with the
purpose written onByteChannelRW[F|S] methods on FSWriteIO one can get an
instance of NIOSeekableByteChannelIO in a typesafe fashion. Since
there's no way to test an already open SeekableByteChannel for write
permission I've NOT provided an IsoMutableRORW, just an IsoMutable to
the read only variant. If you want to force it new up an
NIOByteChannelWriteIO or NIOSeekableByteChannelIO yourself.