Use LowPriorityImplicits for Pure/Catch/ThunkWrap
cats.effect.Sync can delay thunks and we want this behavior in all
wrappers defined for it... even just PureWrap and CatchWrap. Easy to
handle by having ThunkWrap extend PureWrap and CatchWrap. BUT we want to
define a CatchWrap for F[_] which have ApplicativeError but NOT Sync.
Similarly we want to allow PureWrap for any F[_] with an Applicative but
not an ApplicativeError.

Slam dunk case of implicit prioritization. Still had to look it up.
Thanks Stackoverflow!