Add call site demonstration/test example
This is a prototype for how I hope to demonstrate combining algebras
into complex abstract operations and how to then call the resulting
highly generic operations. I've paired a calculation on the FSPath
algebra with a call that compiles into cats.effect.IO which I then wrap
in a ScalaTest test method that confirms the output.

Relevant includes for the call site are made inside the test method
making it clear what's needed to spin the whole thing up.

Next up I'll compile this same test into an Id F[_] to show another
angle. Then I'll write up the same operation and tests in Scalaz 7.2.x.