Add f-bound typing to buffer algebras
All the buffer algebras needed to restrict inputs to the putBuffer
operation to aligned buffer interpreters (in other words IO instances
that wrap a real buffer in the JVM case, or Free implementations) and
need to have aligned, typed results. Other algebras that take buffers
as inputs will restrict parameters to ensure alignment, so these
buffers need to expose enough typing to prove alignment.

Did that by adding type parameters representing the exact module type
with which an IO instance aligns. For all but bytes that meant its
local Module. ByteBuffers can be viewed as all other types though, so
I needed to create a new parent Module type.

All the typing here relies on the projection operator (#) which means
the IO types can't be dependent. They have to be determinable just
from a Module type, not a live instance.