Modify approach to polymophic algebra for buffers
In the branch I'd put bounds on the alignment type parameter, forcing
it to be the Module. In the case of ByteBuffer it's the whole Buffer
system Module. For the others it's their type specific Module. The
bound allows the algebra definitions in the IO traits to return full
leaf types. The trouble is it really complicated the Lift/Term
declarations as the bounds had to be recreated to ensure soundness.
That was gonna be UGLY and the pay off was exposing leaf types from
interpreters which probably isn't even a good idea.

In this commit we allow the alignment parameter to be unrestrained.
This prevents type projection into the Module to define f-bound return
types from operations that accept and return interpreter instances.
That's fine as we have the algebra trait types, with their alignment
parameters. In this commit all the f-bound operations are modified to
use the abstract algebra trait type and the alignment type.

This still uses the alignment type to prevent a FreeByteBufferIO
interpreter from being passed into the JVM interpreter's .read method
for instance. It just doesn't give up the actual runtime type of the
returned interpreter, which is probably better anyway.