Switch to single purpose alignment types
When I tried to carry unrestricted alignment types to the FileIO and
FSIO interfaces I ran into the wall because those interfaces relied on
the alignment type being bound to a Module type to align the multiple
algebras it depends on.

I then realized that we don't need really fine grained alignment, we
just need to ensure that Free interpreters are passed Free
interpreters and Java backed interpreters are passed Java backed
interpreters. There won't be 6 competing and incompatible Java backed
interpreters, ever. So instead of silently using the JVMBufferIOModule
type to align Buffers (and then being out of luck when we get to
FileIO/FSIO) I created a new single purpose JVMA type in the root of
the Java API package.

Any interpreter which implements a wrapped Java library's algebra by
deferring directly to the Java implementation can be tagged with this
dedicated alignment type.