Finish moving whole libraary to single purpose alignment types
FSPathLogic took a heavy reworking to take the IO[_[_], _] shape. I
removed the operation to create a Path from a String because that was
the only one that really needed FSPathLogic to wrap a FileSystem
object. Right now the examples are cheesed together using the
IsoImmutableUnsafe instance for FSPath, which doesn't create FSPath.P
instances with correct D domain labling types.

The better way to create FSPath.P instances is to use the (not yet
created) FileSystem algebra. That's for the future though.

Other than that this commit is a ton of nuking Modules as alignment
types and adding JVMA and FreeA. And gutting Modules. I'm glad I never
cleaned up the Modules. I suspect they'll mostly go away with the new
alignment types approach.