Make some progress fixing up FreeBufferIO
Why'd I stop? ByteBufferIO.putBuffer(ByteBufferReadIO) pretty much
breaks what we're doing. Term is configured to be polymorphic in the
target interpreter's alignment parameter, which is fine. Trouble is
putBuffer, and many other operations in the FileIO algebras that
require input parameters that are aligned interpreters. Free requires
the inputs be stored which happens BEFORE the polymorphic call to
Term.select.

I'm not sure what I'm going to do.